+++ packages/ai/core/generate-text/tool-call-repair-function.ts
@@ -1,5 +1,5 @@
 import { JSONSchema7, LanguageModelV2ToolCall } from '@ai-sdk/provider';
-import { InvalidToolArgumentsError } from '../../src/error/invalid-tool-arguments-error';
+import { InvalidToolInputError } from '../../src/error/invalid-tool-input-error';
 import { NoSuchToolError } from '../../src/error/no-such-tool-error';
 import { ModelMessage } from '../prompt';
 import { ToolSet } from './tool-set';
@@ -14,7 +14,7 @@ import { ToolSet } from './tool-set';
  * @param options.messages - The messages in the current generation step.
  * @param options.toolCall - The tool call that failed to parse.
  * @param options.tools - The tools that are available.
- * @param options.parameterSchema - A function that returns the JSON Schema for a tool.
+ * @param options.inputSchema - A function that returns the JSON Schema for a tool.
  * @param options.error - The error that occurred while parsing the tool call.
  */
 export type ToolCallRepairFunction<TOOLS extends ToolSet> = (options: {
@@ -22,6 +22,6 @@ export type ToolCallRepairFunction<TOOLS extends ToolSet> = (options: {
   messages: ModelMessage[];
   toolCall: LanguageModelV2ToolCall;
   tools: TOOLS;
-  parameterSchema: (options: { toolName: string }) => JSONSchema7;
-  error: NoSuchToolError | InvalidToolArgumentsError;
+  inputSchema: (options: { toolName: string }) => JSONSchema7;
+  error: NoSuchToolError | InvalidToolInputError;
 }) => Promise<LanguageModelV2ToolCall | null>;