+++ packages/ai/src/error/index.ts
@@ -14,7 +14,7 @@ export {
 
 export { InvalidArgumentError } from './invalid-argument-error';
 export { InvalidStreamPartError } from './invalid-stream-part-error';
-export { InvalidToolArgumentsError } from './invalid-tool-arguments-error';
+export { InvalidToolInputError } from './invalid-tool-input-error';
 export { MCPClientError } from './mcp-client-error';
 export { NoImageGeneratedError } from './no-image-generated-error';
 export { NoObjectGeneratedError } from './no-object-generated-error';