+++ packages/anthropic/src/anthropic-messages-language-model.test.ts
@@ -282,7 +282,7 @@ describe('AnthropicMessagesLanguageModel', () => {
           {
             type: 'function',
             name: 'test-tool',
-            parameters: {
+            inputSchema: {
               type: 'object',
               properties: { value: { type: 'string' } },
               required: ['value'],
@@ -303,7 +303,7 @@ describe('AnthropicMessagesLanguageModel', () => {
             "type": "text",
           },
           {
-            "args": "{"value":"example value"}",
+            "input": "{"value":"example value"}",
             "toolCallId": "toolu_1",
             "toolCallType": "function",
             "toolName": "test-tool",
@@ -428,7 +428,7 @@ describe('AnthropicMessagesLanguageModel', () => {
           {
             type: 'function',
             name: 'test-tool',
-            parameters: {
+            inputSchema: {
               type: 'object',
               properties: { value: { type: 'string' } },
               required: ['value'],
@@ -1331,7 +1331,7 @@ describe('AnthropicMessagesLanguageModel', () => {
           {
             type: 'function',
             name: 'test-tool',
-            parameters: {
+            inputSchema: {
               type: 'object',
               properties: { value: { type: 'string' } },
               required: ['value'],
@@ -1364,49 +1364,49 @@ describe('AnthropicMessagesLanguageModel', () => {
             "type": "text",
           },
           {
-            "argsTextDelta": "",
+            "inputTextDelta": "",
             "toolCallId": "toolu_01DBsB4vvYLnBDzZ5rBSxSLs",
             "toolCallType": "function",
             "toolName": "test-tool",
             "type": "tool-call-delta",
           },
           {
-            "argsTextDelta": "{"value",
+            "inputTextDelta": "{"value",
             "toolCallId": "toolu_01DBsB4vvYLnBDzZ5rBSxSLs",
             "toolCallType": "function",
             "toolName": "test-tool",
             "type": "tool-call-delta",
           },
           {
-            "argsTextDelta": "":",
+            "inputTextDelta": "":",
             "toolCallId": "toolu_01DBsB4vvYLnBDzZ5rBSxSLs",
             "toolCallType": "function",
             "toolName": "test-tool",
             "type": "tool-call-delta",
           },
           {
-            "argsTextDelta": ""Spark",
+            "inputTextDelta": ""Spark",
             "toolCallId": "toolu_01DBsB4vvYLnBDzZ5rBSxSLs",
             "toolCallType": "function",
             "toolName": "test-tool",
             "type": "tool-call-delta",
           },
           {
-            "argsTextDelta": "le",
+            "inputTextDelta": "le",
             "toolCallId": "toolu_01DBsB4vvYLnBDzZ5rBSxSLs",
             "toolCallType": "function",
             "toolName": "test-tool",
             "type": "tool-call-delta",
           },
           {
-            "argsTextDelta": " Day"}",
+            "inputTextDelta": " Day"}",
             "toolCallId": "toolu_01DBsB4vvYLnBDzZ5rBSxSLs",
             "toolCallType": "function",
             "toolName": "test-tool",
             "type": "tool-call-delta",
           },
           {
-            "args": "{"value":"Sparkle Day"}",
+            "input": "{"value":"Sparkle Day"}",
             "toolCallId": "toolu_01DBsB4vvYLnBDzZ5rBSxSLs",
             "toolCallType": "function",
             "toolName": "test-tool",