+++ packages/groq/src/groq-chat-language-model.ts
@@ -199,7 +199,7 @@ export class GroqChatLanguageModel implements LanguageModelV2 {
           toolCallType: 'function',
           toolCallId: toolCall.id ?? generateId(),
           toolName: toolCall.function.name,
-          args: toolCall.function.arguments!,
+          input: toolCall.function.arguments!,
         });
       }
     }
@@ -388,7 +388,7 @@ export class GroqChatLanguageModel implements LanguageModelV2 {
                         toolCallType: 'function',
                         toolCallId: toolCall.id,
                         toolName: toolCall.function.name,
-                        argsTextDelta: toolCall.function.arguments,
+                        inputTextDelta: toolCall.function.arguments,
                       });
                     }
 
@@ -400,7 +400,7 @@ export class GroqChatLanguageModel implements LanguageModelV2 {
                         toolCallType: 'function',
                         toolCallId: toolCall.id ?? generateId(),
                         toolName: toolCall.function.name,
-                        args: toolCall.function.arguments,
+                        input: toolCall.function.arguments,
                       });
                       toolCall.hasFinished = true;
                     }
@@ -427,7 +427,7 @@ export class GroqChatLanguageModel implements LanguageModelV2 {
                   toolCallType: 'function',
                   toolCallId: toolCall.id,
                   toolName: toolCall.function.name,
-                  argsTextDelta: toolCallDelta.function.arguments ?? '',
+                  inputTextDelta: toolCallDelta.function.arguments ?? '',
                 });
 
                 // check if tool call is complete
@@ -441,7 +441,7 @@ export class GroqChatLanguageModel implements LanguageModelV2 {
                     toolCallType: 'function',
                     toolCallId: toolCall.id ?? generateId(),
                     toolName: toolCall.function.name,
-                    args: toolCall.function.arguments,
+                    input: toolCall.function.arguments,
                   });
                   toolCall.hasFinished = true;
                 }