+++ packages/openai/src/openai-chat-language-model.ts
@@ -328,7 +328,7 @@ export class OpenAIChatLanguageModel implements LanguageModelV2 {
         toolCallType: 'function',
         toolCallId: toolCall.id ?? generateId(),
         toolName: toolCall.function.name,
-        args: toolCall.function.arguments!,
+        input: toolCall.function.arguments!,
       });
     }
 
@@ -556,7 +556,7 @@ export class OpenAIChatLanguageModel implements LanguageModelV2 {
                         toolCallType: 'function',
                         toolCallId: toolCall.id,
                         toolName: toolCall.function.name,
-                        argsTextDelta: toolCall.function.arguments,
+                        inputTextDelta: toolCall.function.arguments,
                       });
                     }
 
@@ -568,7 +568,7 @@ export class OpenAIChatLanguageModel implements LanguageModelV2 {
                         toolCallType: 'function',
                         toolCallId: toolCall.id ?? generateId(),
                         toolName: toolCall.function.name,
-                        args: toolCall.function.arguments,
+                        input: toolCall.function.arguments,
                       });
                       toolCall.hasFinished = true;
                     }
@@ -595,7 +595,7 @@ export class OpenAIChatLanguageModel implements LanguageModelV2 {
                   toolCallType: 'function',
                   toolCallId: toolCall.id,
                   toolName: toolCall.function.name,
-                  argsTextDelta: toolCallDelta.function.arguments ?? '',
+                  inputTextDelta: toolCallDelta.function.arguments ?? '',
                 });
 
                 // check if tool call is complete
@@ -609,7 +609,7 @@ export class OpenAIChatLanguageModel implements LanguageModelV2 {
                     toolCallType: 'function',
                     toolCallId: toolCall.id ?? generateId(),
                     toolName: toolCall.function.name,
-                    args: toolCall.function.arguments,
+                    input: toolCall.function.arguments,
                   });
                   toolCall.hasFinished = true;
                 }