+++ packages/react/src/use-chat.ui.test.tsx
@@ -625,7 +625,7 @@ describe('onToolCall', () => {
         await toolCallPromise;
         return `test-tool-response: ${toolCall.toolName} ${
           toolCall.toolCallId
-        } ${JSON.stringify(toolCall.args)}`;
+        } ${JSON.stringify(toolCall.input)}`;
       },
     });
 
@@ -664,10 +664,10 @@ describe('onToolCall', () => {
       type: 'stream-chunks',
       chunks: [
         formatStreamPart({
-          type: 'tool-call',
+          type: 'tool-input-available',
           toolCallId: 'tool-call-0',
           toolName: 'test-tool',
-          args: { testArg: 'test-value' },
+          input: { testArg: 'test-value' },
         }),
       ],
     };
@@ -678,8 +678,8 @@ describe('onToolCall', () => {
     expect(
       JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
     ).toStrictEqual({
-      state: 'call',
-      args: { testArg: 'test-value' },
+      state: 'input-available',
+      input: { testArg: 'test-value' },
       toolCallId: 'tool-call-0',
       type: 'tool-test-tool',
     });
@@ -690,11 +690,11 @@ describe('onToolCall', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'result',
-        args: { testArg: 'test-value' },
+        state: 'output-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        result:
+        output:
           'test-tool-response: test-tool tool-call-0 {"testArg":"test-value"}',
       });
     });
@@ -718,13 +718,13 @@ describe('tool invocations', () => {
                   <div data-testid={`tool-invocation-${toolIdx}`}>
                     {JSON.stringify(toolPart)}
                   </div>
-                  {toolPart.state === 'call' && (
+                  {toolPart.state === 'input-available' && (
                     <button
                       data-testid={`add-result-${toolIdx}`}
                       onClick={() => {
                         addToolResult({
                           toolCallId: toolPart.toolCallId,
-                          result: 'test-result',
+                          output: 'test-result',
                         });
                       }}
                     />
@@ -768,7 +768,7 @@ describe('tool invocations', () => {
 
     controller.write(
       formatStreamPart({
-        type: 'tool-call-streaming-start',
+        type: 'tool-input-start',
         toolCallId: 'tool-call-0',
         toolName: 'test-tool',
       }),
@@ -778,7 +778,7 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'partial-call',
+        state: 'input-streaming',
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
       });
@@ -786,9 +786,9 @@ describe('tool invocations', () => {
 
     controller.write(
       formatStreamPart({
-        type: 'tool-call-delta',
+        type: 'tool-input-delta',
         toolCallId: 'tool-call-0',
-        argsTextDelta: '{"testArg":"t',
+        inputTextDelta: '{"testArg":"t',
       }),
     );
 
@@ -796,18 +796,18 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'partial-call',
+        state: 'input-streaming',
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        args: { testArg: 't' },
+        input: { testArg: 't' },
       });
     });
 
     controller.write(
       formatStreamPart({
-        type: 'tool-call-delta',
+        type: 'tool-input-delta',
         toolCallId: 'tool-call-0',
-        argsTextDelta: 'est-value"}}',
+        inputTextDelta: 'est-value"}}',
       }),
     );
 
@@ -815,19 +815,19 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'partial-call',
+        state: 'input-streaming',
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        args: { testArg: 'test-value' },
+        input: { testArg: 'test-value' },
       });
     });
 
     controller.write(
       formatStreamPart({
-        type: 'tool-call',
+        type: 'tool-input-available',
         toolCallId: 'tool-call-0',
         toolName: 'test-tool',
-        args: { testArg: 'test-value' },
+        input: { testArg: 'test-value' },
       }),
     );
 
@@ -835,8 +835,8 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'call',
-        args: { testArg: 'test-value' },
+        state: 'input-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
       });
@@ -844,9 +844,9 @@ describe('tool invocations', () => {
 
     controller.write(
       formatStreamPart({
-        type: 'tool-result',
+        type: 'tool-output-available',
         toolCallId: 'tool-call-0',
-        result: 'test-result',
+        output: 'test-result',
       }),
     );
     controller.close();
@@ -855,11 +855,11 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'result',
-        args: { testArg: 'test-value' },
+        state: 'output-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        result: 'test-result',
+        output: 'test-result',
       });
     });
   });
@@ -875,10 +875,10 @@ describe('tool invocations', () => {
 
     controller.write(
       formatStreamPart({
-        type: 'tool-call',
+        type: 'tool-input-available',
         toolCallId: 'tool-call-0',
         toolName: 'test-tool',
-        args: { testArg: 'test-value' },
+        input: { testArg: 'test-value' },
       }),
     );
 
@@ -886,8 +886,8 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'call',
-        args: { testArg: 'test-value' },
+        state: 'input-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
       });
@@ -895,9 +895,9 @@ describe('tool invocations', () => {
 
     controller.write(
       formatStreamPart({
-        type: 'tool-result',
+        type: 'tool-output-available',
         toolCallId: 'tool-call-0',
-        result: 'test-result',
+        output: 'test-result',
       }),
     );
     controller.close();
@@ -906,11 +906,11 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'result',
-        args: { testArg: 'test-value' },
+        state: 'output-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        result: 'test-result',
+        output: 'test-result',
       });
     });
   });
@@ -928,10 +928,10 @@ describe('tool invocations', () => {
     controller.write(formatStreamPart({ type: 'start-step' }));
     controller.write(
       formatStreamPart({
-        type: 'tool-call',
+        type: 'tool-input-available',
         toolCallId: 'tool-call-0',
         toolName: 'test-tool',
-        args: { testArg: 'test-value' },
+        input: { testArg: 'test-value' },
       }),
     );
 
@@ -939,8 +939,8 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'call',
-        args: { testArg: 'test-value' },
+        state: 'input-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
       });
@@ -952,11 +952,11 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'result',
-        args: { testArg: 'test-value' },
+        state: 'output-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        result: 'test-result',
+        output: 'test-result',
       });
     });
 
@@ -991,9 +991,9 @@ describe('tool invocations', () => {
             {
               type: 'tool-test-tool',
               toolCallId: 'tool-call-0',
-              args: { testArg: 'test-value' },
-              result: 'test-result',
-              state: 'result',
+              input: { testArg: 'test-value' },
+              output: 'test-result',
+              state: 'output-available',
             },
             {
               text: 'more text',
@@ -1022,10 +1022,10 @@ describe('tool invocations', () => {
     // tool call
     controller1.write(
       formatStreamPart({
-        type: 'tool-call',
+        type: 'tool-input-available',
         toolCallId: 'tool-call-0',
         toolName: 'test-tool',
-        args: { testArg: 'test-value' },
+        input: { testArg: 'test-value' },
       }),
     );
 
@@ -1033,8 +1033,8 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'call',
-        args: { testArg: 'test-value' },
+        state: 'input-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
       });
@@ -1048,11 +1048,11 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'result',
-        args: { testArg: 'test-value' },
+        state: 'output-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        result: 'test-result',
+        output: 'test-result',
       });
     });
 
@@ -1089,10 +1089,10 @@ describe('tool invocations', () => {
     // tool call
     controller1.write(
       formatStreamPart({
-        type: 'tool-call',
+        type: 'tool-input-available',
         toolCallId: 'tool-call-0',
         toolName: 'test-tool',
-        args: { testArg: 'test-value' },
+        input: { testArg: 'test-value' },
       }),
     );
 
@@ -1105,8 +1105,8 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'call',
-        args: { testArg: 'test-value' },
+        state: 'input-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
       });
@@ -1120,11 +1120,11 @@ describe('tool invocations', () => {
       expect(
         JSON.parse(screen.getByTestId('message-1').textContent ?? ''),
       ).toStrictEqual({
-        state: 'result',
-        args: { testArg: 'test-value' },
+        state: 'output-available',
+        input: { testArg: 'test-value' },
         toolCallId: 'tool-call-0',
         type: 'tool-test-tool',
-        result: 'test-result',
+        output: 'test-result',
       });
     });
 
@@ -1161,7 +1161,7 @@ describe('maxSteps', () => {
 
           return `test-tool-response: ${toolCall.toolName} ${
             toolCall.toolCallId
-          } ${JSON.stringify(toolCall.args)}`;
+          } ${JSON.stringify(toolCall.input)}`;
         },
         generateId: mockId(),
         maxSteps: 5,
@@ -1197,10 +1197,10 @@ describe('maxSteps', () => {
           type: 'stream-chunks',
           chunks: [
             formatStreamPart({
-              type: 'tool-call',
+              type: 'tool-input-available',
               toolCallId: 'tool-call-0',
               toolName: 'test-tool',
-              args: { testArg: 'test-value' },
+              input: { testArg: 'test-value' },
             }),
           ],
         },
@@ -1228,7 +1228,7 @@ describe('maxSteps', () => {
           onToolCallCounter++;
           return `test-tool-response: ${toolCall.toolName} ${
             toolCall.toolCallId
-          } ${JSON.stringify(toolCall.args)}`;
+          } ${JSON.stringify(toolCall.input)}`;
         },
         generateId: mockId(),
         maxSteps: 5,
@@ -1241,9 +1241,9 @@ describe('maxSteps', () => {
           {messages.map((m, idx) => (
             <div data-testid={`message-${idx}`} key={m.id}>
               {m.parts.filter(isToolUIPart).map((toolPart, toolIdx) =>
-                toolPart.state === 'result' ? (
+                toolPart.state === 'output-available' ? (
                   <div key={toolIdx} data-testid={`tool-${toolIdx}`}>
-                    {JSON.stringify(toolPart.result)}
+                    {JSON.stringify(toolPart.output)}
                   </div>
                 ) : null,
               )}
@@ -1270,10 +1270,10 @@ describe('maxSteps', () => {
           type: 'stream-chunks',
           chunks: [
             formatStreamPart({
-              type: 'tool-call',
+              type: 'tool-input-available',
               toolCallId: 'tool-call-0',
               toolName: 'test-tool',
-              args: { testArg: 'test-value' },
+              input: { testArg: 'test-value' },
             }),
           ],
         },