+++ packages/react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/react
 
+## 2.0.0-alpha.14
+
+### Patch Changes
+
+- 995baa0: fix (react): structuredClone message in replaceMessage
+- Updated dependencies [63f9e9b]
+  - ai@5.0.0-alpha.14
+  - @ai-sdk/provider-utils@3.0.0-alpha.14
+
 ## 2.0.0-alpha.13
 
 ### Major Changes